add history section

+++ controllers/UserController.py
@@ -5,7 +5,7 @@ from servers.UserService import (createUser, updateUser, getUser, getAllUsers, g
                                  getAllUsersByName, searchUsers, deleteUser, deBanneUser, banneUser, disActiveUser,
                                  reActiveUser)
 
-from controllers.AdminController import oauth2_scheme
+from controllers.AdminController import getCurrentAdmin
 
 from repostories.UserRP import ArgumentError, MailExistsError, UserNotFoundError
 
@@ -15,7 +15,10 @@ userRouter = APIRouter(prefix='/users')
 # CRUD APP
 
 @userRouter.post('/')
-def addUserC(new_user: AddUserBaseModel) -> UserResponseBaseModel:
+def addUserC(new_user: AddUserBaseModel, admin=Depends(getCurrentAdmin)) -> UserResponseBaseModel:
+    if admin.authority > 1:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Request Not Allowed")
+
     try:
         return createUser(new_user)  # may generate Error MailExists
     except MailExistsError:
@@ -23,7 +26,10 @@ def addUserC(new_user: AddUserBaseModel) -> UserResponseBaseModel:
 
 
 @userRouter.put('/')
-def updateUserC(model_to_updated: UserUpdateBaseModel) -> UserResponseBaseModel:
+def updateUserC(model_to_updated: UserUpdateBaseModel, admin=Depends(getCurrentAdmin)) -> UserResponseBaseModel:
+    if admin.authority > 1:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Request Not Allowed")
+
     try:
         return updateUser(model_to_updated)
     except (ArgumentError, UserNotFoundError) as e:
@@ -31,7 +37,9 @@ def updateUserC(model_to_updated: UserUpdateBaseModel) -> UserResponseBaseModel:
 
 
 @userRouter.delete('/')
-def deleteUserByIdC(id_: int):
+def deleteUserByIdC(id_: int, admin=Depends(getCurrentAdmin)):
+    if admin.authority > 1:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Request Not Allowed")
     try:
         return deleteUser(id_)
     except UserNotFoundError as e:
@@ -39,7 +47,10 @@ def deleteUserByIdC(id_: int):
 
 
 @userRouter.patch('/password')
-def updatePasswordC(model_password_to_update: UserPasswordModel):
+def updatePasswordC(model_password_to_update: UserPasswordModel, admin=Depends(getCurrentAdmin)):
+    if admin.authority > 1:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Request Not Allowed")
+
     try:
         return updatePasswordUser(model_password_to_update)
     except ArgumentError as e:
@@ -47,7 +58,7 @@ def updatePasswordC(model_password_to_update: UserPasswordModel):
 
 
 @userRouter.get('/{id_}')
-def getUserByIdC(id_: int):
+def getUserByIdC(id_: int, admin=Depends(getCurrentAdmin)):
     try:
         return getUser(id_)
     except UserNotFoundError as e:
@@ -55,7 +66,7 @@ def getUserByIdC(id_: int):
 
 
 @userRouter.get('/mail/{mail}')
-def getUserByMailC(mail: str):
+def getUserByMailC(mail: str, admin=Depends(getCurrentAdmin)):
     try:
         return getUser(mail)
     except UserNotFoundError as e:
@@ -65,46 +76,52 @@ def getUserByMailC(mail: str):
 # GET APP
 
 @userRouter.get('/get/all')
-def getAllUsersC():
+def getAllUsersC(admin=Depends(getCurrentAdmin)):
     return getAllUsers()
 
 
 @userRouter.get('/all/banned')
-def getAllUsersBannedC():
+def getAllUsersBannedC(admin=Depends(getCurrentAdmin)):
     return getAllBannedUsers()
 
 
 @userRouter.get('/all/active')
-def getAllUsersActiveC():
+def getAllUsersActiveC(admin=Depends(getCurrentAdmin)):
     return getAllActiveUsers()
 
 
 @userRouter.get('/all/name/{name}')
-def getAllUsersByNameC(name: str):
+def getAllUsersByNameC(name: str, admin=Depends(getCurrentAdmin)):
     return getAllUsersByName(name)
 
 
 @userRouter.get('/all/search/{name}')
-def searchByNameC(name: str):
+def searchByNameC(name: str, admin=Depends(getCurrentAdmin)):
     return searchUsers(name)
 
 
 # ACTIVE SECTION
 
 @userRouter.post('/user/{id_}')
-def diActiveUserC(id_):
+def diActiveUserC(id_, admin=Depends(getCurrentAdmin)):
+    if admin.authority > 1:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Request Not Allowed")
     return disActiveUser(id_)
 
 
 @userRouter.post('/re-user/{id_}')
-def activeUserC(id_: int):
+def activeUserC(id_: int, admin=Depends(getCurrentAdmin)):
+    if admin.authority > 1:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Request Not Allowed")
     return reActiveUser(id_)
 
 
 # BANNE SECTION
 
 @userRouter.post('/banne/{id_}')
-def banneUserC(id_: int):
+def banneUserC(id_: int, admin=Depends(getCurrentAdmin)):
+    if admin.authority > 1:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Request Not Allowed")
     try:
         return banneUser(id_)
     except ArgumentError as e:
@@ -112,7 +129,9 @@ def banneUserC(id_: int):
 
 
 @userRouter.post('/de-banne/{id_}')
-def deBanneUserC(id_: int):
+def deBanneUserC(id_: int, admin=Depends(getCurrentAdmin)):
+    if admin.authority > 1:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Request Not Allowed")
     try:
         return deBanneUser(id_)
     except ArgumentError as e:
@@ -120,5 +139,5 @@ def deBanneUserC(id_: int):
 
 
 @userRouter.get('/count/all')
-def getNumberOfUsersC(current_user=Depends(oauth2_scheme)):
+def getNumberOfUsersC(current_user=Depends(getCurrentAdmin)):
     return getNumberOfUsers()
